used send grid to send emails

+++ src/api/v1/auth/routes.js
@@ -2,7 +2,7 @@ const express = require("express");
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
 const prisma = require("../../../../prisma/prismaClient");
-const { sendOtpMail } = require("../../../helper/mailer");
+const sendOtpEmail = require("../../../helper/mailer");
 const { jwtSecret, otpExpireMinutes } = require("../../../config");
 const authRouter = express.Router();
 
@@ -14,22 +14,32 @@ const generateExpiresAt = () =>
 
 authRouter.post("/signup", async (req, res) => {
   try {
-    const { email, password, confirmPassword } = req.body;
-    if (!email || !password || password !== confirmPassword) {
-      return res.status(400).json({ message: "Invalid credentials" });
+    const { fullName, email, password, confirmPassword } = req.body;
+    if (!email || !password || !fullName || password !== confirmPassword) {
+      return res
+        .status(400)
+        .json({ status: "failure", message: "Invalid credentials" });
     }
 
     const existing = await prisma.user.findUnique({ where: { email } });
     if (existing && existing.isVerified)
-      return res.status(400).json({ message: "Email exists" });
+      return res
+        .status(400)
+        .json({ status: "failure", message: "Email exists" });
 
     // delete the unverified user as they can not access the system anyways
-    if (existing && !existing.isVerified)
+    if (existing && !existing.isVerified) {
+      await prisma.otpToken.deleteMany({
+        where: {
+          userId: existing.id,
+        },
+      });
       await prisma.user.delete({
         where: {
           email: email,
         },
       });
+    }
 
     const hashed = await bcrypt.hash(password, 10);
     const user = await prisma.user.create({
@@ -45,20 +55,33 @@ authRouter.post("/signup", async (req, res) => {
     });
 
     await prisma.otpToken.upsert({
-      data: {
+      where: {
+        userId_purpose: {
+          userId: user.id,
+          purpose: "SIGNUP",
+        },
+      },
+      create: {
         userId: user.id,
         otpCode: otp,
         createdAt: new Date(),
-        expiresAt,
+        expiresAt: expiresAt,
         purpose: "SIGNUP",
+      },
+      update: {
+        otpCode: otp,
+        createdAt: new Date(),
+        expiresAt: expiresAt,
       },
     });
 
-    await sendOtpMail(email, otp);
-    return res.status(200).json({ message: "OTP sent" });
+    await sendOtpEmail(email, otp);
+    return res.status(200).json({ status: "success", message: "OTP sent" });
   } catch (err) {
     console.error(err);
-    res.status(500).json({ message: "Something went wrong" });
+    res
+      .status(500)
+      .json({ status: "failure", message: "Something went wrong" });
   }
 });
 
@@ -91,8 +114,13 @@ authRouter.post("/verify-otp", async (req, res) => {
     const jwtToken = jwt.sign({ id: user.id, email: user.email }, jwtSecret, {
       expiresIn: "7d",
     });
-
-    res.status(200).json({ message: "OTP verified", token: jwtToken });
+    res.status(200).cookie("token", jwtToken, {
+      httpOnly: true,
+      secure: process.env.NODE_ENV === "production",
+      sameSite: "Lax",
+      maxAge: 7 * 24 * 60 * 60 * 1000,
+    });
+    // res.status(200).json({ message: "OTP verified", token: jwtToken });
   } catch (err) {
     console.error(err);
     res.status(500).json({ message: "Something went wrong" });
@@ -115,8 +143,13 @@ authRouter.post("/login", async (req, res) => {
     const jwtToken = jwt.sign({ id: user.id, email: user.email }, jwtSecret, {
       expiresIn: "7d",
     });
-
-    res.status(200).json({ token: jwtToken });
+    res.status(200).cookie("token", jwtToken, {
+      httpOnly: true,
+      secure: process.env.NODE_ENV === "production",
+      sameSite: "Lax",
+      maxAge: 7 * 24 * 60 * 60 * 1000,
+    });
+    // res.status(200).json({ token: jwtToken });
   } catch (err) {
     console.error(err);
     res.status(500).json({ message: "Something went wrong" });
@@ -185,7 +218,7 @@ authRouter.post("/resend-otp", async (req, res) => {
     });
 
     try {
-      await sendOtpMail(email, otp);
+      await sendOtpEmail(email, otp);
       return res.status(200).json({ message: "OTP resent to your email." });
     } catch (error) {
       console.error("Email send failed:", error);
@@ -198,7 +231,6 @@ authRouter.post("/resend-otp", async (req, res) => {
     res.status(500).json({ message: "Something went wrong" });
   }
 });
-
 
 authRouter.get("/me", async (req, res) => {
   try {
